Introduce abstract KMS client
