+++ simplyblock_core/kms/_exceptions.py
@@ -1,0 +1,3 @@
+class KMSException(Exception):
+    def __init__(self, message):
+        self.message = message
